interception of .tor TLDs

+++ clean.sh
@@ -6,9 +6,13 @@ rm -rf build/
 #remove libscrypt build files
 ((cd src/libs/libscrypt-1.20/ && make clean))
 
-#clean up tor-client
-rm -rf tor-client
-mv tor-client-orig tor-client
+#clean up built binaries
 rm tor OnioNS_d
+
+# if tor-client backup exists, restore from it
+if [ -d "tor-client-orig" ]; then
+    rm -rf tor-client
+    mv tor-client-orig tor-client
+fi
 
 echo "Successfully removed the build directory."
